refactor(reorganizing): break up geocoding package into multiple files
AFFECTS PACKAGES:
@esri/arcgis-rest-geocoder

ISSUES CLOSED: #216
