+++ packages/arcgis-rest-geocoder/src/index.ts
@@ -1,1 +1,4 @@
-export * from "./geocoder";
+export * from "./geocode";
+export * from "./suggest";
+export * from "./reverse";
+export * from "./bulk";
